.model-reference/requirements.txt
asgiref==3.7.2
cloudinary==1.41.0
cryptography==41.0.7
dj-database-url==2.1.0
dj-rest-auth==5.0.2
Django==4.2.11
django-allauth==0.57.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
gunicorn==21.2.0
oauthlib==3.2.2
Pillow==10.2.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2024.1
requests-oauthlib==1.3.1
setuptools==69.5.1
sqlparse==0.4.4
urllib3==2.2.1
whitenoise==6.6.0
